Create index.html
Original template updated to include basic text about SUCS to get the page going.
In the Carousel, see the updated logo graphic (x3). The previous version lost quality where the spark (star thing) was cut out. The diver has been filled in and a better quality star has been cut out again. The text below has less white shadow now. In the second version, the spark is left white over the diver, and in version three the white spark is bigger and in the background. We need to pick a graphic to at least use for now.
The Modal Login and Register links and Footer info were also updated.